requirements.txt
amqp==2.3.2
billiard==3.5.0.4
celery==4.2.1
certifi==2018.10.15
chardet==3.0.4
Django==2.1.1
django-bootstrap4==0.0.7
django-celery-beat==1.2.0
django-celery-results==1.0.1
django-ckeditor==5.6.1
django-js-asset==1.1.0
django-paypal==0.5.0
django-timezone-field==3.0
djangorestframework==3.8.2
djangorestframework-simplejwt==3.2.3
idna==2.7
kombu==4.2.1
mysqlclient==1.3.13
Pillow==5.2.0
PyJWT==1.6.4
python-crontab==2.3.5
python-dateutil==2.7.3
pytz==2018.5
requests==2.20.0
six==1.11.0
SQLAlchemy==1.2.12
stripe==2.10.1
urllib3==1.24
vine==1.1.4
django-debug-toolbar==1.10.1
